push

+++ tests/test_socket_factory.cpp
@@ -113,7 +113,7 @@ int main() {
 
         auto clt_result = SocketFactory::createTcpClient(AddressFamily::IPv4,
             ConnectArgs{"127.0.0.1",
-                Port{static_cast<uint16_t>(srvPort.load())},
+                Port{srvPort.load()},
                 Milliseconds{2000}});
         REQUIRE(clt_result.isSuccess());
         auto& client = clt_result.value();
